requirements.txt
numpy==1.20.3
pandas==1.3.0
pytorch==1.9.0
torchvision==0.10.0
cudatoolkit=10.2
seaborn==0.11.1
pytorch-lightning==1.4.0
timm==0.4.12
scikit-learn==0.24.2
